requirements.txt
# MR. HEALTH Data Platform -- Python Dependencies

# Data generation
faker>=22.0.0
pandas>=2.0.0

# GCP SDKs
google-cloud-storage>=2.14.0
google-cloud-bigquery>=3.17.0
functions-framework>=3.5.0

# Data processing
pyarrow>=14.0.0

# Configuration
pyyaml>=6.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# PostgreSQL integration
psycopg2-binary>=2.9.0
paramiko>=3.4.0
google-cloud-secret-manager>=2.18.0

# Code quality
ruff>=0.3.0

# Airflow (for local DAG development and testing)
apache-airflow==2.8.1
apache-airflow-providers-google>=10.12.0
